Moved fancy forms into vis editor
