+++ src/legacy/ui/public/vis/editors/default/agg.js
@@ -21,7 +21,7 @@ import './agg_params';
 import './agg_add';
 import { Direction } from './keyboard_move';
 import _ from 'lodash';
-import '../../../fancy_forms';
+import './fancy_forms';
 import { uiModules } from '../../../modules';
 import aggTemplate from './agg.html';
 import { move } from '../../../utils/collection';
