+++ src/legacy/ui/public/vis/editors/default/fancy_forms/__tests__/fancy_forms.js
