+++ src/legacy/ui/public/vis/editors/default/fancy_forms/__tests__/nested_fancy_forms.js
