+++ src/legacy/ui/public/vis/editors/default/fancy_forms/fancy_forms.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { uiModules } from '../modules';
+import { uiModules } from '../../../../modules';
 
 import { decorateFormController } from './kbn_form_controller';
 import { decorateModelController } from './kbn_model_controller';
