+++ src/legacy/ui/public/vis/editors/default/fancy_forms/index.js
